Changed few minor things in the simulation configuration and simulation panels.
Finished the abstract class for the PersistenceManager and implemented
an InMemoryPersistenceManager which keeps now all the data required in
memory and provides methods to access and modify it.

Changed all the required code to reference this new PersistenceManager
instead of the old references to the MainGui local variables.

Solved few bugs with the clues and threats table and entry duplicates.